Added timer in a seperate component
