+++ src/app/components/goal-dashboard-child-components/start-timer/start-timer.component.html
@@ -1,0 +1,5 @@
+<div class="timer">
+    <h3>Time left to complete following program!</h3>
+    <p>{{days}} days, {{hours}} hours, {{minutes}} minutes, {{seconds}} seconds, </p>
+    <p>{{message}}</p>
+</div>
