+++ src/app/components/goal-dashboard/goal-dashboard.component.ts
@@ -1,4 +1,4 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, OnInit, ViewChild } from '@angular/core';
 import { Router } from '@angular/router';
 import * as $ from 'jquery';
 import { GoalDashbordService } from 'src/app/services/goal-dashbord.service';
@@ -8,6 +8,7 @@ import { Workout } from 'src/app/models/workout.model';
 import { WorkoutPageService } from 'src/app/services/workout-page.service';
 import { ProgrammePageService } from 'src/app/services/programme-page.service';
 import { Programme } from 'src/app/models/programme.model';
+import { StartTimerComponent } from '../goal-dashboard-child-components/start-timer/start-timer.component';
 
 
 @Component({
@@ -16,6 +17,9 @@ import { Programme } from 'src/app/models/programme.model';
   styleUrls: ['./goal-dashboard.component.css']
 })
 export class GoalDashboardComponent implements OnInit {
+
+  //---start-timer link
+  @ViewChild(StartTimerComponent, {static : true}) startTimer : StartTimerComponent | undefined;
 
   username: string = JSON.parse(sessionStorage.getItem("current-user")||'{}').firstName
   
@@ -69,7 +73,7 @@ export class GoalDashboardComponent implements OnInit {
   program_ex_map_name_in7 = new Map();
 
   //---Commits logged an commits finished
-  dayCommit = new Map();
+  dayCommit = new Map(); 
   dayCommitInitial = new Map();
   
   //---workout id
@@ -116,7 +120,6 @@ export class GoalDashboardComponent implements OnInit {
 
   ngOnInit(): void {
   
-
     //---get weekly submitted progress from local storage
    
     if(localStorage.getItem("weekly_finish")){
@@ -205,7 +208,7 @@ export class GoalDashboardComponent implements OnInit {
 
       // //---start timer
       this.countDownDate = Number(localStorage.getItem("countdown_timer"))
-      this.startTimer()
+      this.callTimer(this.countDownDate)
       
   }
     
@@ -263,8 +266,22 @@ export class GoalDashboardComponent implements OnInit {
   daily_progress(value: number){
     this.dailyProgress=value
   }
+  daily_progress_display(value: number){
+    this.dailyProgress_display=value
+  }
   weekly_progress(value: number){
     this.weeklyProgress=value
+  }
+  sumExerciseInitial(value:number){
+    this.sumExInitial = value
+  }
+
+  //---commits
+  dayCommits(value: any){
+    this.dayCommit = value
+  }
+  dayCommitsInitial(value: any){
+    this.dayCommitInitial = value
   }
 
 
@@ -319,7 +336,7 @@ export class GoalDashboardComponent implements OnInit {
      }else{
       this.countDownDate = new Date().setDate(new Date().getDate()+7);
      }
-     this.startTimer();
+     this.callTimer(this.countDownDate);
     }
  
   //---go to details page
@@ -355,196 +372,9 @@ export class GoalDashboardComponent implements OnInit {
     location.reload();
   }
 
-  startTimer(){
-  //---Update the count down every 1 second
-  const sevenDayTimer = setInterval(() => {
+  callTimer(value:number){
+    this.startTimer?.startTimer(value)
+  }
 
-    //---disable buttons and menues
-    this.disable();
-
-    //---Get today's date and time
-    let now = new Date().getTime();
-    
-    //---calculate difference between end date and now
-    let difference = this.countDownDate - now;
-
-     //---Calculations for days, hours, minutes and seconds
-    this.days = Math.floor(difference / (1000 * 60 * 60 * 24));
-    this.hours = Math.floor((difference % (1000 * 60 * 60 * 24)) / (1000 * 60 * 60))+1;
-    this.minutes = Math.floor((difference % (1000 * 60 * 60)) / (1000 * 60));
-    this.seconds = Math.floor((difference % (1000 * 60)) / 1000);
-
-    //---commit end date to localstorage
-    localStorage.setItem("countdown_timer", JSON.stringify(this.countDownDate))
-
-    //---Switch to different day commits every daily cycle
-   //---for day 1
-   if (this.days===6 && this.hours>0 && this.minutes>0 && this.seconds>0){
-     
-     this.dayCommit = this.program_ex_map_name
-     this.dayCommitInitial = new Map(this.program_ex_map_name_in)
-     
-     this.dayOne = this.oneDate.getFullYear()+"-"+(this.oneDate.getMonth()+1)+"-"+this.oneDate.getDate()
-     this.dayTwo = this.twoDate.getFullYear()+"-"+(this.twoDate.getMonth()+1)+"-"+this.twoDate.getDate()
-     this.dayThree = this.threeDate.getFullYear()+"-"+(this.threeDate.getMonth()+1)+"-"+this.threeDate.getDate()
-     this.dayFour = this.fourDate.getFullYear()+"-"+(this.fourDate.getMonth()+1)+"-"+this.fourDate.getDate()
-     this.dayFive = this.fiveDate.getFullYear()+"-"+(this.fiveDate.getMonth()+1)+"-"+this.fiveDate.getDate()
-     this.daySix = this.sixDate.getFullYear()+"-"+(this.sixDate.getMonth()+1)+"-"+this.sixDate.getDate()
-     this.daySeven = this.sevenDate.getFullYear()+"-"+(this.sevenDate.getMonth()+1)+"-"+this.sevenDate.getDate()
-          
-    }
-   //---for day 2
-   else if(this.days===5 && this.hours>0 && this.minutes>0 && this.seconds>0){
-
-     if(this.days===5 && this.hours===23 && this.minutes===59 && this.seconds===59){
-       localStorage.removeItem("daily_finish")
-       this.dailyProgress = 0;
-       this.dailyProgress_display= '';
-       this.sumExInitial=0;
-     }
-
-     this.dayCommit = this.program_ex_map_name2
-     this.dayCommitInitial = new Map(this.program_ex_map_name_in2)
-
-     this.dayOne = this.mtwoDate.getFullYear()+"-"+(this.mtwoDate.getMonth()+1)+"-"+this.mtwoDate.getDate()
-     this.dayTwo = this.oneDate.getFullYear()+"-"+(this.oneDate.getMonth()+1)+"-"+this.oneDate.getDate()
-     this.dayThree = this.twoDate.getFullYear()+"-"+(this.twoDate.getMonth()+1)+"-"+this.twoDate.getDate()
-     this.dayFour = this.threeDate.getFullYear()+"-"+(this.threeDate.getMonth()+1)+"-"+this.threeDate.getDate()
-     this.dayFive = this.fourDate.getFullYear()+"-"+(this.fourDate.getMonth()+1)+"-"+this.fourDate.getDate()
-     this.daySix = this.fiveDate.getFullYear()+"-"+(this.fiveDate.getMonth()+1)+"-"+this.fiveDate.getDate()
-     this.daySeven = this.sixDate.getFullYear()+"-"+(this.sixDate.getMonth()+1)+"-"+this.sixDate.getDate()
-    }
-
-    //---for day 3
-    else if(this.days===4 && this.hours>0 && this.minutes>0 && this.seconds>0){
-
-     if(this.days===4 && this.hours===23 && this.minutes===59 && this.seconds===59){
-       localStorage.removeItem("daily_finish")
-       this.dailyProgress = 0;
-       this.dailyProgress_display= '';
-       this.sumExInitial=0;
-     }
-
-     this.dayCommit = this.program_ex_map_name3
-     this.dayCommitInitial = new Map(this.program_ex_map_name_in3)
-
-     this.dayOne = this.mthreeDate.getFullYear()+"-"+(this.mthreeDate.getMonth()+1)+"-"+this.mthreeDate.getDate()
-     this.dayTwo = this.mtwoDate.getFullYear()+"-"+(this.mtwoDate.getMonth()+1)+"-"+this.mtwoDate.getDate()
-     this.dayThree = this.oneDate.getFullYear()+"-"+(this.oneDate.getMonth()+1)+"-"+this.oneDate.getDate()
-     this.dayFour = this.twoDate.getFullYear()+"-"+(this.twoDate.getMonth()+1)+"-"+this.twoDate.getDate()
-     this.dayFive = this.threeDate.getFullYear()+"-"+(this.threeDate.getMonth()+1)+"-"+this.threeDate.getDate()
-     this.daySix = this.fourDate.getFullYear()+"-"+(this.fourDate.getMonth()+1)+"-"+this.fourDate.getDate()
-     this.daySeven = this.fiveDate.getFullYear()+"-"+(this.fiveDate.getMonth()+1)+"-"+this.fiveDate.getDate()
-    }
-
-   //---for day 4
-   else if (this.days===3 && this.hours>0 && this.minutes>0 && this.seconds>0){
-
-     if(this.days===3 && this.hours===23 && this.minutes===59 && this.seconds===59){
-       localStorage.removeItem("daily_finish")
-       this.dailyProgress = 0;
-       this.dailyProgress_display= '';
-       this.sumExInitial=0;
-     }
-
-     this.dayCommit = this.program_ex_map_name4
-     this.dayCommitInitial = new Map(this.program_ex_map_name_in4)
-
-     this.dayOne = this.mfourDate.getFullYear()+"-"+(this.mfourDate.getMonth()+1)+"-"+this.mfourDate.getDate()
-     this.dayTwo = this.mthreeDate.getFullYear()+"-"+(this.mthreeDate.getMonth()+1)+"-"+this.mthreeDate.getDate()
-     this.dayThree = this.mtwoDate.getFullYear()+"-"+(this.mtwoDate.getMonth()+1)+"-"+this.mtwoDate.getDate()
-     this.dayFour = this.oneDate.getFullYear()+"-"+(this.oneDate.getMonth()+1)+"-"+this.oneDate.getDate()
-     this.dayFive = this.twoDate.getFullYear()+"-"+(this.twoDate.getMonth()+1)+"-"+this.twoDate.getDate()
-     this.daySix = this.threeDate.getFullYear()+"-"+(this.threeDate.getMonth()+1)+"-"+this.threeDate.getDate()
-     this.daySeven = this.fourDate.getFullYear()+"-"+(this.fourDate.getMonth()+1)+"-"+this.fourDate.getDate()
-    
-    }
-
-   //---for day 5
-   else if(this.days===2 && this.hours>0 && this.minutes>0 && this.seconds>0){
-
-     if(this.days===2 && this.hours===23 && this.minutes===59 && this.seconds===59){
-       localStorage.removeItem("daily_finish")
-       this.dailyProgress = 0;
-       this.dailyProgress_display= '';
-       this.sumExInitial=0;
-     }
-
-     this.dayCommit = this.program_ex_map_name5
-     this.dayCommitInitial = new Map(this.program_ex_map_name_in5)
-    
-     this.dayOne = this.mfiveDate.getFullYear()+"-"+(this.mfiveDate.getMonth()+1)+"-"+this.mfiveDate.getDate()
-     this.dayTwo = this.mfourDate.getFullYear()+"-"+(this.mfourDate.getMonth()+1)+"-"+this.mfourDate.getDate()
-     this.dayThree = this.mthreeDate.getFullYear()+"-"+(this.mthreeDate.getMonth()+1)+"-"+this.mthreeDate.getDate()
-     this.dayFour = this.mtwoDate.getFullYear()+"-"+(this.mtwoDate.getMonth()+1)+"-"+this.mtwoDate.getDate()
-     this.dayFive = this.oneDate.getFullYear()+"-"+(this.oneDate.getMonth()+1)+"-"+this.oneDate.getDate()
-     this.daySix = this.twoDate.getFullYear()+"-"+(this.twoDate.getMonth()+1)+"-"+this.twoDate.getDate()
-     this.daySeven = this.threeDate.getFullYear()+"-"+(this.threeDate.getMonth()+1)+"-"+this.threeDate.getDate()
-
-
-   }
-
-   //---for day 6
-   else if(this.days===1 && this.hours>0 && this.minutes>0 && this.seconds>0){
-
-     if(this.days===1 && this.hours===23 && this.minutes===59 && this.seconds===59){
-       localStorage.removeItem("daily_finish")
-       this.dailyProgress = 0;
-       this.dailyProgress_display= '';
-       this.sumExInitial=0;
-     }
-
-     this.dayCommit = this.program_ex_map_name6
-     this.dayCommitInitial = new Map(this.program_ex_map_name_in6)
-    
-     this.dayOne = this.msixDate.getFullYear()+"-"+(this.msixDate.getMonth()+1)+"-"+this.msixDate.getDate()
-     this.dayTwo = this.mfiveDate.getFullYear()+"-"+(this.mfiveDate.getMonth()+1)+"-"+this.mfiveDate.getDate()
-     this.dayThree = this.mfourDate.getFullYear()+"-"+(this.mfourDate.getMonth()+1)+"-"+this.mfourDate.getDate()
-     this.dayFour = this.mthreeDate.getFullYear()+"-"+(this.mthreeDate.getMonth()+1)+"-"+this.mthreeDate.getDate()
-     this.dayFive = this.mtwoDate.getFullYear()+"-"+(this.mtwoDate.getMonth()+1)+"-"+this.mtwoDate.getDate()
-     this.daySix = this.oneDate.getFullYear()+"-"+(this.oneDate.getMonth()+1)+"-"+this.oneDate.getDate()
-     this.daySeven = this.twoDate.getFullYear()+"-"+(this.twoDate.getMonth()+1)+"-"+this.twoDate.getDate()
-
-   }
-
-   //---for day 7
-   else if(this.days===0 && this.hours>0 && this.minutes>0 && this.seconds>0){
-
-     if(this.days===0 && this.hours===23 && this.minutes===59 && this.seconds===59){
-       localStorage.removeItem("daily_finish")
-       this.dailyProgress = 0;
-       this.dailyProgress_display= '';
-       this.sumExInitial=0;
-     }
-
-     this.dayCommit = this.program_ex_map_name7
-     this.dayCommitInitial = new Map(this.program_ex_map_name_in7)
-
-     this.dayOne = this.msevenDate.getFullYear()+"-"+(this.msevenDate.getMonth()+1)+"-"+this.msevenDate.getDate()
-     this.dayTwo = this.msixDate.getFullYear()+"-"+(this.msixDate.getMonth()+1)+"-"+this.msixDate.getDate()
-     this.dayThree = this.mfiveDate.getFullYear()+"-"+(this.mfiveDate.getMonth()+1)+"-"+this.mfiveDate.getDate()
-     this.dayFour = this.mfourDate.getFullYear()+"-"+(this.mfourDate.getMonth()+1)+"-"+this.mfourDate.getDate()
-     this.dayFive = this.mthreeDate.getFullYear()+"-"+(this.mthreeDate.getMonth()+1)+"-"+this.mthreeDate.getDate()
-     this.daySix = this.mtwoDate.getFullYear()+"-"+(this.mtwoDate.getMonth()+1)+"-"+this.mtwoDate.getDate()
-     this.daySeven = this.oneDate.getFullYear()+"-"+(this.oneDate.getMonth()+1)+"-"+this.oneDate.getDate()
-   
-   }
-   //---Display message when count down is finished
-   else if(difference === 0) {
-     const message = "Deadline expired";
-     
-     this.uncommitGoal()
-     }
-   }, 1000);
 }
-
- disable(){
- //---disable menues and buttons
- $('.program-options').attr('disabled','disabled');
- $('.workout-options').attr('disabled','disabled');
- $('.exercise-options').attr('disabled','disabled');
- $('.clear').attr('disabled','disabled');
- $('.commit-program').attr('disabled','disabled');
- $('.day-buttons').attr('disabled','disabled');
-}}
